create unit test for part load
created a preliminary unit test for the part load calculations of Vapor compression chillers and moved the previous 'load_distribution' into 'chiller_vapor_compression'
