+++ cea/technologies/chiller_vapor_compression.py
@@ -5,12 +5,12 @@ from __future__ import division
 import pandas as pd
 from math import log, ceil
 import numpy as np
-import cea.technologies.load_distribution as load_distribution
 
 from cea.technologies.constants import G_VALUE_CENTRALIZED, G_VALUE_DECENTRALIZED, CHILLER_DELTA_T_HEX_CT, \
     CHILLER_DELTA_T_APPROACH, T_EVAP_AHU, T_EVAP_ARU, T_EVAP_SCU, DT_NETWORK_CENTRALIZED, CENTRALIZED_AUX_PERCENTAGE, \
-    DECENTRALIZED_AUX_PERCENTAGE
+    DECENTRALIZED_AUX_PERCENTAGE, COMPRESSOR_TYPE_LIMIT_LOW, COMPRESSOR_TYPE_LIMIT_HIGH, ASHRAE_CAPACITY_LIMIT
 from cea.analysis.costs.equations import calc_capex_annualized, calc_opex_annualized
+from cea.utilities.physics import kelvin_to_fahrenheit
 
 __author__ = "Thuy-An Nguyen"
 __copyright__ = "Copyright 2015, Architecture and Building Systems - ETH Zurich"
@@ -72,7 +72,7 @@ def calc_COP(T_cw_in_K, T_chw_re_K, q_chw_load_Wh):
 
 def calc_COP_with_carnot_efficiency(peak_cooling_load, q_chw_load_Wh, T_chw_sup_K, T_cw_in_K, g_value,
                                     min_chiller_size, max_chiller_size, scale):
-    PLF = load_distribution.calc_averaged_PLF(peak_cooling_load, q_chw_load_Wh, T_chw_sup_K, T_cw_in_K,
+    PLF = calc_averaged_PLF(peak_cooling_load, q_chw_load_Wh, T_chw_sup_K, T_cw_in_K,
                                               min_chiller_size, max_chiller_size, scale)  # calculates the weighted average Part load factor across all chillers based on load distribution
     cop_chiller = g_value * T_chw_sup_K / (T_cw_in_K - T_chw_sup_K) * PLF
     return cop_chiller
@@ -186,12 +186,127 @@ def get_max_VCC_unit_size(locator, VCC_code='CH3'):
     max_VCC_unit_size_W = max(VCC_cost_data['cap_max'].values)
     return max_VCC_unit_size_W
 
+def calc_averaged_PLF(peak_cooling_load, q_chw_load_Wh, T_chw_sup_K, T_cw_in_K, min_chiller_size, max_chiller_size, scale):
+    """
+    Calculates the part load factor of installed Vapor compression chillers for a given cooling load.
+    Includes the design of the chillers based on peak load and chiller plant scale to define the part load ratio.
+    :param float peak_cooling_load:
+    :param float q_chw_load_Wh:
+    :param float T_chw_sup_K:
+    :param float T_cw_in_K:
+    :param float min_chiller_size:
+    :param float max_chiller_size:
+    :param str scale: either "BUILDING" or "DISTRICT"
+    :return float averaged_PLF: averaged part load factor over all chillers
+    """
+    design_capacity = peak_cooling_load  # * 1.15 # for future implementation, a safety factor could be introduced. As of now this would be in conflict with the master_to_slave_variables.WS_BaseVCC_size_W
+    if scale == 'BUILDING':
+        if design_capacity <= COMPRESSOR_TYPE_LIMIT_LOW: # according to ASHRAE 90.1 Appendix G: if design cooling load smaller than lower limit, implement one screw chiller
+            source_type = 'WATER'
+            compressor_type = 'SCREW'
+            n_units = 1
+        elif COMPRESSOR_TYPE_LIMIT_LOW < design_capacity < COMPRESSOR_TYPE_LIMIT_HIGH: # according to ASHRAE 90.1 Appendix G: if design cooling load between limits, implement two screw chillers
+            source_type = 'WATER'
+            compressor_type = 'SCREW'
+            n_units = 2
+        elif design_capacity >= COMPRESSOR_TYPE_LIMIT_HIGH: # according to ASHRAE 90.1 Appendix G: if design cooling load larger than upper limit, implement centrifugal chillers
+            source_type = 'WATER'
+            compressor_type = 'CENTRIFUGAL'
+            n_units = ceil(design_capacity / ASHRAE_CAPACITY_LIMIT) # according to ASHRAE 90.1 Appendix G, chiller shall not be large then 800 tons (2813 kW)
+        cooling_capacity_per_unit = design_capacity / n_units  # calculate the capacity per chiller installed
+
+    elif scale == 'DISTRICT':
+        source_type = 'WATER'
+        compressor_type = 'CENTRIFUGAL'
+        if design_capacity <= (2*min_chiller_size): # design one chiller for small scale DCS
+            n_units = 1
+            cooling_capacity_per_unit = max(design_capacity, min_chiller_size)
+        elif (2*min_chiller_size) <= design_capacity <= max_chiller_size: # design two chillers above the twice the minimum chiller size
+            n_units = 2
+            cooling_capacity_per_unit = design_capacity / n_units
+        elif design_capacity >= max_chiller_size: # design a minimum of 2 chillers if above the maximum chiller size
+            n_units = max(2, ceil(design_capacity / max_chiller_size)) # have minimum size of capacity to quailfy as DCS, have minimum of 2 chillers
+            cooling_capacity_per_unit = design_capacity / n_units
+    else:
+        raise AssertionError('No or unexpected scale was assigned even though this should never happen.', scale)
+
+    available_capacity_per_unit = calc_available_capacity(cooling_capacity_per_unit, source_type, compressor_type, T_chw_sup_K, T_cw_in_K) # calculate the available capacity(dependent on conditions)
+
+    ### calculate the load distribution across the chillers heuristically, assuming the PLF factor is monotonously increasing with increasing PLR. Filling one chiller after the other.
+    n_chillers_filled = int(q_chw_load_Wh // available_capacity_per_unit)
+    part_load_chiller = float(divmod(q_chw_load_Wh, available_capacity_per_unit)[1])/ float(available_capacity_per_unit)
+
+    load_distribution_list = []
+    for i in range (n_chillers_filled):
+        load_distribution_list.append(1)
+    load_distribution_list.append(part_load_chiller)
+    for i in range(int(n_units)-n_chillers_filled-1):
+        load_distribution_list.append(0)
+    load_distribution = np.array(load_distribution_list)
+
+    averaged_PLF = np.sum(calc_PLF(load_distribution, source_type, compressor_type) * load_distribution * available_capacity_per_unit) / q_chw_load_Wh # calculates the weighted average PLF value
+    return averaged_PLF
+
+def calc_PLF(PLR, source_type, compressor_type):
+    """
+    takes the part load ratio as an input and outputs the part load factor
+    coefficients taken from https://comnet.org/index.php/382-chillers and only includes water source electrical chillers TODO: create database entry dependent on technology
+    :param np.array PLR: part load ratio for each chiller
+    :param string source_type: only 'WATER' acceptable as of now, can potentially be extended to 'AIR' source.
+    :param string compressor_type: either 'CENTRIFUGAL' or 'SCREW'
+    :return np.array PLF: part load factor for each chiller
+    """
+    if source_type == 'WATER' and compressor_type == 'CENTRIFUGAL':
+        plf_a = 0.17149273
+        plf_b = 0.58820208
+        plf_c = 0.23737257
+    if source_type == 'WATER' and compressor_type == 'SCREW':
+        plf_a = 0.33018833
+        plf_b = 0.23554291
+        plf_c = 0.46070828
+
+    PLF = plf_a + plf_b * PLR + plf_c * PLR ** 2
+    return PLF
+
+def calc_available_capacity(rated_capacity, source_type, compressor_type, T_chw_sup_K, T_cw_in_K):
+    """
+    calculates the available Chiller capacity based on the rated capacity
+    coefficients taken from https://comnet.org/index.php/382-chillers  TODO: create database entry dependent on technology
+    :param float rated_capacity: rated capacity of chiller
+    :param string source_type: only 'WATER' acceptable as of now, can potentially be extended to 'AIR' source.
+    :param string compressor_type: either 'CENTRIFUGAL' or 'SCREW'
+    :return np.array PLF: part load factor for each chiller
+    :param float T_chw_sup_K: supplied chilled water temperature in Kelvin
+    :param float T_cw_in_K: condenser water supply temperature in Kelvin
+    """
+    if source_type == 'WATER' and compressor_type == 'CENTRIFUGAL':
+        q_a = -0.29861976
+        q_b = 0.02996076
+        q_c = -0.00080125
+        q_d = 0.01736268
+        q_e = -0.00032606
+        q_f = 0.00063139
+    if source_type == 'WATER' and compressor_type == 'SCREW':
+        q_a = 0.33269598
+        q_b = 0.00729116
+        q_c = -0.00049938
+        q_d = 0.01598983
+        q_e = -0.00028254
+        q_f = 0.00052346
+
+    t_chws_F = kelvin_to_fahrenheit(T_chw_sup_K)
+    t_cws_F = kelvin_to_fahrenheit(T_cw_in_K)
+
+    available_capacity = rated_capacity * (q_a + q_b*t_chws_F + q_c*t_chws_F**2 + q_d*t_cws_F + q_e * t_cws_F**2 + q_f*t_chws_F*t_cws_F)
+    return available_capacity
+
+
 def main():
     scale = 'DISTRICT'
-    peak_cooling_load = 1500000
-    Qc_W = 1000000
-    max_chiller_size = 3500000
-    min_chiller_size = 1000000
+    peak_cooling_load = 40000000
+    Qc_W = 25000000
+    max_chiller_size = 14000000
+    min_chiller_size = 1758000
     T_chw_sup_K = 273.15 + 6
     T_chw_re_K = 273.15 + 11
     T_cw_in_K = 273.15 + 28
